move status match checker to new package
