+++ backend/gateway/statuserr/statuserr.go
@@ -1,0 +1,19 @@
+package statuserr
+
+import (
+	statuspb "google.golang.org/genproto/googleapis/rpc/status"
+	"google.golang.org/grpc/codes"
+)
+
+func AllStatusMatch(c codes.Code, spb ...*statuspb.Status) bool {
+	if len(spb) == 0 {
+		return false
+	}
+
+	for _, s := range spb {
+		if s.Code != int32(c) {
+			return false
+		}
+	}
+	return true
+}
