+++ backend/module/resolver/response.go
@@ -11,6 +11,7 @@ import (
 	"google.golang.org/grpc/status"
 
 	apiv1 "github.com/lyft/clutch/backend/api/api/v1"
+	"github.com/lyft/clutch/backend/gateway/statuserr"
 	"github.com/lyft/clutch/backend/resolver"
 )
 
@@ -60,19 +61,6 @@ func (r *response) marshalResults(results *resolver.Results) error {
 	return nil
 }
 
-func allStatusMatch(c codes.Code, sl []*statuspb.Status) bool {
-	if len(sl) == 0 {
-		return false
-	}
-
-	for _, s := range sl {
-		if s.Code != int32(c) {
-			return false
-		}
-	}
-	return true
-}
-
 func (r *response) isError(wanted string, searchedSchemas []string) error {
 	// If results, errors will be returned as partial failures.
 	if len(r.Results) > 0 {
@@ -86,7 +74,7 @@ func (r *response) isError(wanted string, searchedSchemas []string) error {
 	// If there were failures and no results we wrap the errors.
 	if len(r.PartialFailures) > 0 {
 		// Use 400 unless all codes were 404.
-		if !allStatusMatch(codes.NotFound, r.PartialFailures) {
+		if !statuserr.AllStatusMatch(codes.NotFound, r.PartialFailures...) {
 			code = codes.FailedPrecondition
 			msg = fmt.Sprintf("one or more errors were encountered searching for '%s'", wanted)
 		}
