+++ backend/module/resolver/response_test.go
@@ -104,55 +104,3 @@ func TestIsError(t *testing.T) {
 		assert.NoError(t, err)
 	}
 }
-
-func TestAllStatusMatch(t *testing.T) {
-	cases := []struct {
-		name   string
-		code   codes.Code
-		expect bool
-		status []*statuspb.Status
-	}{
-		{
-			name:   "empty",
-			code:   codes.NotFound,
-			expect: false,
-		},
-		{
-			name:   "singleMatch",
-			code:   codes.Unavailable,
-			status: []*statuspb.Status{{Code: int32(codes.Unavailable)}},
-			expect: true,
-		},
-		{
-			name:   "multipleMatch",
-			code:   codes.Unauthenticated,
-			status: []*statuspb.Status{{Code: int32(codes.Unauthenticated)}, {Code: int32(codes.Unauthenticated)}},
-			expect: true,
-		},
-		{
-			name:   "singleNoMatch",
-			code:   codes.InvalidArgument,
-			status: []*statuspb.Status{{Code: int32(codes.Unauthenticated)}},
-			expect: false,
-		},
-		{
-			name:   "multipleNoMatch",
-			code:   codes.InvalidArgument,
-			status: []*statuspb.Status{{Code: int32(codes.Unauthenticated)}, {Code: int32(codes.Unimplemented)}},
-			expect: false,
-		},
-		{
-			name:   "multipleSomeMatch",
-			code:   codes.InvalidArgument,
-			status: []*statuspb.Status{{Code: int32(codes.InvalidArgument)}, {Code: int32(codes.Unimplemented)}},
-			expect: false,
-		},
-	}
-
-	for _, c := range cases {
-		c := c
-		t.Run(c.name, func(t *testing.T) {
-			assert.Equal(t, c.expect, allStatusMatch(c.code, c.status))
-		})
-	}
-}
